% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emusauce.R
\name{emusauce}
\alias{emusauce}
\title{Voice quality measures in R}
\usage{
emusauce(
  inputDir,
  pitch = TRUE,
  formant = TRUE,
  bw = TRUE,
  harmonicAmplitude = TRUE,
  slope = TRUE,
  harmonicAmplitudeUncorrected = TRUE,
  slopeUncorrected = TRUE,
  cpp = TRUE,
  hnr = TRUE,
  intensity = TRUE,
  soe = TRUE,
  intervalFixed = 0.005,
  f0min = 50,
  f0max = 300,
  windowLength = 0.025,
  maxNumFormants = 5,
  bw_hawksMiller = TRUE,
  praatsauce_output = NULL
)
}
\arguments{
\item{inputDir}{String giving a directory containing sound files to analyze.}

\item{pitch}{Boolean; should pitch values be returned? Default is `TRUE`.}

\item{formant}{Boolean; should formant values be returned? Default is `TRUE`.}

\item{bw}{Boolean; should formant bandwidths be returned? Default is `TRUE`.}

\item{harmonicAmplitude}{Boolean; should corrected harmonic amplitudes be
returned? Default is `TRUE`.}

\item{slope}{Boolean; should corrected spectral slope values be returned?
Default is `TRUE`.}

\item{harmonicAmplitudeUncorrected}{Boolean; should uncorrected harmonic
amplitudes be returned? Default is `TRUE`.}

\item{slopeUncorrected}{Boolean; should uncorrected spectral slope values
be returned? Default is `TRUE`.}

\item{cpp}{Boolean; should cepstral peak prominence values be returned?
Default is `TRUE`.}

\item{hnr}{Boolean; should harmonics-to-noise ratios be returned? Default
is `TRUE`.}

\item{intensity}{Boolean; should root-mean-squared intensity values be
returned? Default is `TRUE`.}

\item{soe}{Boolean; should strength of excitation values be returned?
Default is `TRUE`.}

\item{intervalFixed}{Numeric; how often should measures be taken (in
seconds)? Default is `0.005`.}

\item{f0min}{Integer giving the pitch floor. Default is `50`.}

\item{f0max}{Integer giving the pitch ceiling. Default is `300`.}

\item{windowLength}{Numeric giving the length of the analysis window (in
seconds). Default is `0.025`.}

\item{maxNumFormants}{Integer giving the maximum number of formants to
estimate. Default is `5`.}

\item{bw_hawksMiller}{Boolean; should bandwidths be estimated using the
Hawks-Miller formula? Default is `TRUE`. If `FALSE`, raw bandwidths are
returned.}

\item{praatsauce_output}{Optional data frame containing existing measures
from PraatSauce.}
}
\value{
Data frame with the specified measures.
}
\description{
Estimates a range of voice quality measures over all sound files in a
directory relying mainly on functions from the EMU-SDMS libraries.
}
\examples{
datapath <- system.file('extdata/audio', package='sauceshelf')
sauce <- emusauce(datapath)
}
