% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bwHawksMiller.R
\name{get_bwHawksMiller}
\alias{get_bwHawksMiller}
\title{Estimate bandwidths using the Hawks-Miller formula}
\usage{
get_bwHawksMiller(f0, fmt)
}
\arguments{
\item{f0}{Numeric vector of pitch values.}

\item{fmt}{Numeric vector of formant values.}
}
\value{
Numeric vector of estimated bandwidths values.
}
\description{
Formula estimation of formant bandwidths based on real pitch and formant
measurements.
}
\examples{
# not now
}
