% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CPP.R
\name{get_CPP}
\alias{get_CPP}
\title{Estimate cepstral peak prominence}
\usage{
get_CPP(
  filelist,
  inputDir,
  windowLength = 0.025,
  f0min = 50,
  f0max = 300,
  intervalFixed = 0.005,
  beginTime = 0,
  praatsauce_output = NULL
)
}
\arguments{
\item{filelist}{Vector of strings giving the names of sound files to analyze.}

\item{inputDir}{String giving the directory where sound files are located.}

\item{windowLength}{Numeric giving the length of the analysis window (in
seconds). Default is `0.025`.}

\item{f0min}{Integer giving the pitch floor. Default is `50`.}

\item{f0max}{Integer giving the pitch ceiling. Default is `300`.}

\item{intervalFixed}{Numeric; how often should measures be taken (in
seconds)? Default is `0.005`.}

\item{beginTime}{Numeric; where should the first measure be taken (in
seconds)? Default is `0`.}

\item{praatsauce_output}{Optional data frame containing existing measures
from PraatSauce. Used to ensure an equal number of rows.}
}
\value{
A data frame with CPP values.
}
\description{
Estimate cepstral peak prominence from a list of sound files by comparing
actual cepstral peak to the corresponding location in
the regression-predicted linear slope of the cepstrum.
}
\examples{
datapath <- system.file('extdata/audio', package='sauceshelf')
fls <- list.files(datapath, pattern='.wav')
cpp <- get_CPP(fls, datapath)
}
