% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formants.R
\name{get_formants}
\alias{get_formants}
\title{Estimate formants and formant bandwidths}
\usage{
get_formants(
  filelist,
  inputDir = ".",
  intervalFixed = 0.005,
  windowLength = 0.025,
  measureBandwidths = FALSE,
  maxNumFormants = 5,
  beginTime = 0,
  praatsauce_output = NULL
)
}
\arguments{
\item{filelist}{Vector of strings giving the names of sound files to analyze.}

\item{inputDir}{String giving the directory where sound files are located.}

\item{intervalFixed}{Numeric; how often should measures be taken (in
seconds)? Default is `0.005`.}

\item{windowLength}{Numeric giving the length of the analysis window (in
seconds). Default is `0.025`.}

\item{measureBandwidths}{Boolean; should bandwidths be returned? Default is
`FALSE`.}

\item{maxNumFormants}{Integer giving the maximum number of formants to
estimate. Default is `5`.}

\item{beginTime}{Numeric; where should the first measure be taken (in
seconds)? Default is `0`.}

\item{praatsauce_output}{Optional data frame containing existing measures
from PraatSauce. Used to ensure an equal number of rows.}
}
\value{
A data frame with formants and optionally bandwidths values.
}
\description{
Estimate formants and optionally bandwidths for a list of sound files.
Wrapper for [wrassp::forest] returning a data frame.
}
\examples{
datapath <- system.file('extdata/audio', package='sauceshelf')
fls <- list.files(datapath, pattern='.wav')
fmt <- get_formants(fls, datapath)
}
