% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intensity.R
\name{get_intensity}
\alias{get_intensity}
\title{Estimate root-mean-squared intensity}
\usage{
get_intensity(
  filelist,
  inputDir,
  intervalFixed = 0.005,
  f0min = 50,
  beginTime = 0,
  praatsauce_output = NULL
)
}
\arguments{
\item{filelist}{Vector of strings giving the names of sound files to analyze.}

\item{inputDir}{String giving the directory where sound files are located.}

\item{intervalFixed}{Numeric; how often should measures be taken (in
seconds)? Default is `0.005`.}

\item{f0min}{Integer giving the pitch floor. Default is `50`.}

\item{beginTime}{Numeric; where should the first measure be taken (in
seconds)? Default is `0`.}

\item{praatsauce_output}{Optional data frame containing existing measures
from PraatSauce. Used to ensure an equal number of rows.}
}
\value{
A data frame with intensity values.
}
\description{
Estimate root-mean-squared intensity of a list of sounds files.
Wrapper for [wrassp::rmsana] returning a data frame.
}
\examples{
datapath <- system.file('extdata/audio', package='sauceshelf')
fls <- list.files(datapath, pattern='.wav')
rms <- get_intensity(fls, datapath)
}
