% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectralMeasures.R
\name{get_spectralMeasures}
\alias{get_spectralMeasures}
\title{Estimate harmonic amplitudes and spectral slope measures}
\usage{
get_spectralMeasures(
  filelist,
  inputDir,
  f0,
  fmt,
  bw,
  intervalFixed = 0.005,
  windowLength = 0.025,
  slope = TRUE,
  slopeUncorrected = TRUE,
  beginTime = 0,
  praatsauce_output = NULL
)
}
\arguments{
\item{filelist}{Vector of strings giving the names of sound files to analyze.}

\item{inputDir}{String giving the directory where sound files are located.}

\item{f0}{Data frame consisting of at least a column `file` with sound file
names, a column `t` with time in seconds, and a column `f0` with pitch
measures.}

\item{fmt}{Data frame consisting of at least a column `file` with sound file
names, a column `t` with time in seconds, and columns `F1`, `F2`, and `F3`
with formant measurements.}

\item{bw}{Data frame consisting of at least a column `file` with sound file
names, a column `t` with time in seconds, and columns `B1`, `B2`, and `B3`
with formant bandiwdths estimations.}

\item{intervalFixed}{Numeric; how often should measures be taken (in
seconds)? Default is `0.005`.}

\item{windowLength}{Numeric giving the length of the analysis window (in
seconds). Default is `0.025`.}

\item{slope}{Boolean; should spectral slope measures be returned? Default
is `TRUE`.}

\item{slopeUncorrected}{Boolean; should uncorrected spectral slope measures
be returned? Default is `TRUE`.}

\item{beginTime}{Numeric; where should the first measure be taken (in
seconds)? Default is `0`.}

\item{praatsauce_output}{Optional data frame containing existing measures
from PraatSauce. Used to ensure an equal number of rows.}
}
\value{
A data frame with harmonic amplitudes and spectral slope measures.
}
\description{
Based on provided pitch and formant measures, this function estimates a
commonly used range of harmonic amplitude and spectral slope measures
internally in R. Spectra are generated with Blackman windows using
[wrassp::dftSpectrum].
}
\examples{
# not now
}
