% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sauce.R
\name{load_sauce}
\alias{load_sauce}
\title{Load voice quality measurements made in PraatSauce}
\usage{
load_sauce(fileLoc, useTextGrid = FALSE, emuDB = FALSE)
}
\arguments{
\item{fileLoc}{String giving the location of PraatSauce output file}

\item{useTextGrid}{Logical; what was the value of the `useTextGrid`
parameter in the parameters file when calling PraatSauce? Default is `FALSE`.}

\item{emuDB}{Logical; is the audio data structured as an EMU database?
Default is `FALSE`.}
}
\value{
Formatted data frame.
}
\description{
Load tab-separated file of voice quality measures made
by running the PraatSauce scripts and properly format data frame
}
\examples{
datapath <- system.file('extdata/audio/out.tsv', package='sauceshelf')
sauce <- sauceshelf::load_sauce(datapath)
}
