% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praatsauce_load.R
\name{praatsauce_load}
\alias{praatsauce_load}
\title{Load voice quality measurements made in PraatSauce}
\usage{
praatsauce_load(fileLoc, useTextGrid = FALSE, emuDB = FALSE, na_output = NA)
}
\arguments{
\item{fileLoc}{String giving the location of PraatSauce output file}

\item{useTextGrid}{Logical; what was the value of the `useTextGrid`
parameter in the parameters file when calling PraatSauce? Default is `FALSE`.}

\item{emuDB}{Logical; is the audio data structured as an EMU database?
Default is `FALSE`.}

\item{na_output}{How should infelicitous values be coded? Default is `NA`.}
}
\value{
Formatted data frame based on output from PraatSauce.
}
\description{
Load tab-separated file of voice quality measures made
by running the PraatSauce scripts and properly format data frame
}
\examples{
datapath <- system.file('extdata/text/out.tsv', package='sauceshelf')
sauce <- praatsauce_load(datapath)
}
