% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaper_install.R
\name{reaper_install}
\alias{reaper_install}
\title{Install REAPER}
\usage{
reaper_install()
}
\value{
Used for side effects.
}
\description{
Install David Talkin's Robust Epoch and Pitch Estimator program as part
of the `sauceshelf` R package.
}
\examples{
\dontrun{
reaper_install()
}
}
