% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauce2ssff.R
\name{sauce2ssff}
\alias{sauce2ssff}
\title{Convert sauce measures to SSFF files}
\usage{
sauce2ssff(emuDBhandle, sauce)
}
\arguments{
\item{emuDBhandle}{Handle of a loaded EMU database.}

\item{sauce}{Data frame containing sauce measures. This should be generated
by running e.g. `praatsauce` on the same loaded EMU database.}
}
\value{
Used for side effects.
}
\description{
Converts sauce measures to SSFF files and adds them to a loaded EMU database.
}
\examples{
\donttest{
# Create demo data and load demo database
emuR::create_emuRdemoData(tempdir())
db_path <- paste0(tempdir(), '/emuR_demoData/ae_emuDB')
db <- emuR::load_emuDB(db_path)

# Which SSFF tracks are available?
emuR <- emuR::list_ssffTrackDefinitions(db)

# Get sauce
sauce <- praatsauce(db)
sauce2ssff(db, sauce)

# Did it work?
emuR::list_ssffTrackDefinitions(db)
}
}
