% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauce_hirst2pass.R
\name{sauce_hirst2pass}
\alias{sauce_hirst2pass}
\title{Voice quality measures with by-speaker two-pass pitch estimation}
\usage{
sauce_hirst2pass(
  inputDir,
  sauceFunction = praatsauce,
  recursive = FALSE,
  firstPass_f0min = 50,
  firstPass_f0max = 700,
  min_multiplier = 0.75,
  max_multiplier = 1.5,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{inputDir}{String giving the location of a directory with WAV files
to process. Note that this should be the entire file path; regular
R style path extension will not work, as the directory should  be
relative to the location of the Praat script. This should be a nested
directory, where each directory contains the sound files of a speaker.
Alternatively, this can be the handle of a loaded EMU database.}

\item{sauceFunction}{Which function should be called to estimate voice
quality measures? Should be either `praatsauce` (default) or `mixedsauce`,
but currently only works properly for `praatsauce`.}

\item{recursive}{Logical; should sound files in subsubdirectories of
`inputDir` be analyzed? Default is `FALSE`.}

\item{firstPass_f0min}{Integer giving the pitch floor to be used for the
first pass in Hz. Default is `50`.}

\item{firstPass_f0max}{Integer giving the pitch ceiling to be used for the
first pass in Hz. Default is `700`.}

\item{min_multiplier}{Numeric; the pitch floor in the second pass corresponds
to the first quartile from the first pass multiplied by this value.
Default is `0.75`.}

\item{max_multiplier}{Numeric; the pitch ceiling in the second pass
corresponds to the third quartile from the first pass multiplied by this
value.  Default is `1.5`.}

\item{verbose}{Logical; should a message be printed to the console whenever
a speaker has been processed? Default is `TRUE`.}

\item{...}{Further arguments passed on to the function specified by
`sauceFunction`.}
}
\value{
Data frame containing results of chosen measures.
}
\description{
Estimates a range of voice quality measures using a two-pass procedure for
estimating suitable by-speaker pitch floor and ceiling values,
following e.g. De Looze & Hirst (2021). Pitch is estimated for each speaker
using liberal floor and ceiling values, and suitable estimates are then
computed from the distribution of these values.
}
\examples{
# NOT NOW
}
