% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voicesauce_load.R
\name{voicesauce_load}
\alias{voicesauce_load}
\title{Load and format output from VoiceSauce}
\usage{
voicesauce_load(fileLoc, normalize_output = TRUE, na_output = NA)
}
\arguments{
\item{fileLoc}{String pointing to the output file from VoiceSauce}

\item{normalize_output}{Boolean; should the column names and contents from
VoiceSauce be normalized to look like the output of [emusauce] and
[praatsauce]? Default is `TRUE`.}

\item{na_output}{How should infelicitous values be coded? Default is `NA`.}
}
\value{
A data frame with rows and columns corresponding to the VoiceSauce
output.
}
\description{
Loads an output file with measures calculated in the VoiceSauce program
into R.
}
\examples{
datapath <- system.file('extdata/text', package='sauceshelf')
sauce <- voicesauce_load(paste0(datapath, '/output.txt'))
}
