% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauce2ssff.R
\name{sauce2ssff}
\alias{sauce2ssff}
\title{Convert sauce measures to SSFF files}
\usage{
sauce2ssff(emuDBhandle, sauce)
}
\arguments{
\item{emuDBhandle}{Handle of a loaded EMU database.}

\item{sauce}{Data frame containing sauce measures. This should be generated
by running e.g. `praatsauce` on the same loaded EMU database.}
}
\value{
Used for side effects.
}
\description{
Converts sauce measures to SSFF files and adds them to a loaded EMU database.
}
\examples{
# not right now
}
